v1.0.3 Release

+++ Model/Webhook/DeleteAllWebhooks.php
@@ -3,10 +3,13 @@ declare(strict_types=1);
 
 namespace TreviPay\TreviPayMagento\Model\Webhook;
 
+use Exception;
 use Magento\Store\Model\ScopeInterface;
 use TreviPay\TreviPay\Exception\ApiClientException;
 use TreviPay\TreviPayMagento\Model\TreviPayFactory;
 use TreviPay\TreviPayMagento\Model\ConfigProvider;
+use TreviPay\TreviPay\Exception\ResponseException;
+use Psr\Log\LoggerInterface;
 
 class DeleteAllWebhooks
 {
@@ -20,12 +23,19 @@ class DeleteAllWebhooks
      */
     private $treviPayFactory;
 
+    /**
+     * @var LoggerInterface
+     */
+    private $logger;
+
     public function __construct(
         ConfigProvider $configProvider,
-        TreviPayFactory $treviPayFactory
+        TreviPayFactory $treviPayFactory,
+        LoggerInterface $logger
     ) {
         $this->configProvider = $configProvider;
         $this->treviPayFactory = $treviPayFactory;
+        $this->logger = $logger;
     }
 
     /**
@@ -40,16 +50,26 @@ class DeleteAllWebhooks
             return;
         }
 
+        if (isset($webhooks['id'])) {
+            $webhooks = [($webhooks)];
+        }
+
+        $baseUrl = $this->configProvider->getBaseUrl($scope, $scopeId);
         $apiKey = $this->configProvider->getApiKeyForCreatedWebhooks($scope, $scopeId);
         $treviPay = $this->treviPayFactory->create([], $scope, $scopeId);
+
         foreach ($webhooks as $index => $webhook) {
-            if (is_array($webhook)) {
-                if (isset($webhook['id'])) {
+            try {
+                if (is_array($webhook) && (isset($webhook['id'])) && strpos($webhook['webhook_url'], $baseUrl) === 0) {
                     $treviPay->webhooks->delete($webhook['id'], $apiKey);
                 }
-            } elseif ($index === 'id') {
-                $treviPay->webhooks->delete($webhook, $apiKey);
-                break;
+
+            } catch (ResponseException $e) {
+                if ($e->getCode() === 404) {
+                    continue;
+                }
+
+                throw $e;
             }
         }
     }
